Planning: used float in dp_poly_path and dp_st_graph.

+++ modules/planning/tasks/dp_st_speed/dp_st_cost.h
@@ -43,52 +43,52 @@ class DpStCost {
                     const std::vector<const PathObstacle*>& obstacles,
                     const common::TrajectoryPoint& init_point);
 
-  double GetObstacleCost(const StGraphPoint& point);
+  float GetObstacleCost(const StGraphPoint& point);
 
-  double GetReferenceCost(const STPoint& point,
+  float GetReferenceCost(const STPoint& point,
                           const STPoint& reference_point) const;
 
-  double GetSpeedCost(const STPoint& first, const STPoint& second,
-                      const double speed_limit) const;
+  float GetSpeedCost(const STPoint& first, const STPoint& second,
+                      const float speed_limit) const;
 
-  double GetAccelCostByTwoPoints(const double pre_speed, const STPoint& first,
+  float GetAccelCostByTwoPoints(const float pre_speed, const STPoint& first,
                                  const STPoint& second);
-  double GetAccelCostByThreePoints(const STPoint& first, const STPoint& second,
+  float GetAccelCostByThreePoints(const STPoint& first, const STPoint& second,
                                    const STPoint& third);
 
-  double GetJerkCostByTwoPoints(const double pre_speed, const double pre_acc,
+  float GetJerkCostByTwoPoints(const float pre_speed, const float pre_acc,
                                 const STPoint& pre_point,
                                 const STPoint& curr_point);
-  double GetJerkCostByThreePoints(const double first_speed,
+  float GetJerkCostByThreePoints(const float first_speed,
                                   const STPoint& first_point,
                                   const STPoint& second_point,
                                   const STPoint& third_point);
 
-  double GetJerkCostByFourPoints(const STPoint& first, const STPoint& second,
+  float GetJerkCostByFourPoints(const STPoint& first, const STPoint& second,
                                  const STPoint& third, const STPoint& fourth);
 
  private:
-  double GetAccelCost(const double accel);
-  double JerkCost(const double jerk);
+  float GetAccelCost(const float accel);
+  float JerkCost(const float jerk);
 
   void AddToKeepClearRange(const std::vector<const PathObstacle*>& obstacles);
   static void SortAndMergeRange(
-      std::vector<std::pair<double, double>>* keep_clear_range_);
-  bool InKeepClearRange(double s) const;
+      std::vector<std::pair<float, float>>* keep_clear_range_);
+  bool InKeepClearRange(float s) const;
 
   const DpStSpeedConfig& config_;
   const std::vector<const PathObstacle*>& obstacles_;
   const common::TrajectoryPoint& init_point_;
 
-  double unit_t_ = 0.0;
+  float unit_t_ = 0.0;
 
   std::unordered_map<std::string, int> boundary_map_;
-  std::vector<std::vector<std::pair<double, double>>> boundary_cost_;
+  std::vector<std::vector<std::pair<float, float>>> boundary_cost_;
 
-  std::vector<std::pair<double, double>> keep_clear_range_;
+  std::vector<std::pair<float, float>> keep_clear_range_;
 
-  std::array<double, 200> accel_cost_;
-  std::array<double, 400> jerk_cost_;
+  std::array<float, 200> accel_cost_;
+  std::array<float, 400> jerk_cost_;
 };
 
 }  // namespace planning
